refactor: facade와 module 서비스 일부 수정
NotificationFacade에서 수행되던 기능을 EmitterService로 이동

resolves: #102
