+++ src/main/java/com/zerobase/foodlier/module/notification/repository/notification/NotificationRepository.java
@@ -1,4 +1,4 @@
-package com.zerobase.foodlier.module.notification.repository;
+package com.zerobase.foodlier.module.notification.repository.notification;
 
 import com.zerobase.foodlier.module.notification.domain.model.Notification;
 import com.zerobase.foodlier.module.notification.dto.NotificationDto;
@@ -11,7 +11,7 @@ import org.springframework.data.repository.query.Param;
 public interface NotificationRepository extends JpaRepository<Notification, Long> {
 
     @Query("select new com.zerobase.foodlier.module.notification.dto.NotificationDto(" +
-           "n.id, n.content, n.notificationType, n.sendAt, n.isRead) " +
+           "n.id, n.content, n.notificationType, n.sendAt, n.isRead, n.url) " +
             "from  Notification n join n.member m " +
             "where m.id = :memberId " +
             "order by n.sendAt desc ")
