+++ src/main/java/com/zerobase/foodlier/module/notification/service/emitter/EmitterService.java
@@ -9,9 +9,11 @@ import java.util.Map;
 public interface EmitterService {
     String makeTimeIncludeId(String email);
     SseEmitter createEmitter(String userEmail);
-    Map<String, Object> findMissingNotification(String userEmail);
+    Map<String, Object> findEventCaches(String userEmail);
     void deleteEmitter(String emitterId);
     void send(SseEmitter emitter, String eventId, String emitterId, Object data);
     void createEventCache(String emitterId, Notification notification);
     Map<String, SseEmitter> findAllEmitter(String receiverEmail);
+    boolean isEmitterExists(Map<String, SseEmitter> emitterMap);
+    boolean hasLostData(String lastEventId);
 }
