+++ src/main/java/com/zerobase/foodlier/module/notification/service/notification/NotificationServiceImpl.java
@@ -4,7 +4,7 @@ import com.zerobase.foodlier.module.member.member.domain.model.Member;
 import com.zerobase.foodlier.module.notification.domain.model.Notification;
 import com.zerobase.foodlier.module.notification.domain.type.NotificationType;
 import com.zerobase.foodlier.module.notification.dto.NotificationDto;
-import com.zerobase.foodlier.module.notification.repository.NotificationRepository;
+import com.zerobase.foodlier.module.notification.repository.notification.NotificationRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
